requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
tenacity==8.5.0
orjson==3.10.7
pytest==8.3.3
pytest-asyncio==0.24.0
